+++ app/[locale]/projects/page.js
@@ -4,12 +4,13 @@ import SummaryCard from "@/components/cards/SummaryCard";
 import { Box } from "@mui/material";
 import Link from "next/link";
 import ProjectOfTheWeek from "@/components/cards/ProjectOfTheWeek";
-import { FundContext } from "@/context/FundContext";
+
 import CategoryFiltering from "@/components/category/CategoryFiltering";
 import LoaderStyle from "@/components/helper/LoaderStyle";
+import { useAuth } from "@/context/AuthContext";
 
 const Home = () => {
-  const { projects } = useContext(FundContext);
+  const { projects } = useAuth();
   const [data, setData] = useState(projects ?? []);
   const [projectOfWeek, setProjectOFWeek] = useState("");
   useEffect(() => {
@@ -19,6 +20,7 @@ const Home = () => {
   // Find project with maximum contributions
   useEffect(() => {
     let max = 0;
+    console.log(data)
     if (data.length > 0) {
       data.forEach((project) => {
         if (project.raised > max) {
