+++ app/config/firebase.js
@@ -13,8 +13,8 @@ const firebaseConfig = {
 };
 
 const app = initializeApp(firebaseConfig)
-export const auth = getAuth(app)
 
+export const auth = getAuth(app)
 export const  db = getFirestore(app);
